feat: Handle unknown OpTypes in BackendInfo
If Nexus or a similar service is reporting OpTypes that are unknown to
the local pytket version, we are currently raising an attribute error
when trying to convert to a pytket BackendInfo.

This is not necessarily a fatal issue in most places as the user may
want to access results or similar data and does not care about the
BackendInfo. This commit changes the behavior to just raise a warning
instead if the OpType does not have that attribute.